+++ src/main/java/com/maxmind/minfraud/response/CreditCard.java
@@ -1,31 +1,38 @@
 package com.maxmind.minfraud.response;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonValue;
 
 /**
  * This class contains minFraud response data related to the credit card.
  */
 public final class CreditCard {
     private final Issuer issuer;
+    private final String brand;
     private final String country;
     private final Boolean isIssuedInBillingAddressCountry;
     private final Boolean isPrepaid;
+    private final String type;
 
     public CreditCard(
+            @JsonProperty("brand") String brand,
             @JsonProperty("country") String country,
             @JsonProperty("is_issued_in_billing_address_country") Boolean isIssuedInBillingAddressCountry,
             @JsonProperty("is_prepaid") Boolean isPrepaid,
-            @JsonProperty("issuer") Issuer issuer
+            @JsonProperty("issuer") Issuer issuer,
+            @JsonProperty("type") String type
     ) {
+        this.brand = brand;
         this.country = country;
         this.isIssuedInBillingAddressCountry = isIssuedInBillingAddressCountry;
         this.isPrepaid = isPrepaid;
         this.issuer = issuer == null ? new Issuer() : issuer;
+        this.type = type;
     }
 
     public CreditCard() {
-        this(null, null, null, null);
+        this(null, null, null, null, null, null);
     }
 
     /**
@@ -33,6 +40,13 @@ public final class CreditCard {
      */
     public Issuer getIssuer() {
         return issuer;
+    }
+
+    /**
+     * @return The credit card brand.
+     */
+    public String getBrand() {
+        return brand;
     }
 
     /**
@@ -67,13 +81,22 @@ public final class CreditCard {
         return isPrepaid;
     }
 
+    /**
+     * @return The credit card type.
+     */
+    public String getType() {
+        return type;
+    }
+
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder("CreditCard{");
         sb.append("issuer=").append(this.issuer);
+        sb.append(", brand='").append(this.brand).append('\'');
         sb.append(", country='").append(this.country).append('\'');
         sb.append(", isIssuedInBillingAddressCountry=").append(this.isIssuedInBillingAddressCountry);
         sb.append(", isPrepaid=").append(this.isPrepaid);
+        sb.append(", type='").append(this.type).append('\'');
         sb.append('}');
         return sb.toString();
     }