+++ src/main/java/com/maxmind/minfraud/response/InsightsResponse.java
@@ -3,6 +3,7 @@ package com.maxmind.minfraud.response;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.List;
+import java.util.UUID;
 
 /**
  * This class provides a model for the minFraud Insights response.
@@ -10,14 +11,18 @@ import java.util.List;
 public final class InsightsResponse extends ScoreResponse {
     private final IpAddress ipAddress;
     private final CreditCard creditCard;
+    private final Device device;
+    private final Email email;
     private final ShippingAddress shippingAddress;
     private final BillingAddress billingAddress;
 
     public InsightsResponse(
             @JsonProperty("billing_address") BillingAddress billingAddress,
             @JsonProperty("credit_card")  CreditCard creditCard,
+            @JsonProperty("device")  Device device,
+            @JsonProperty("email")  Email email,
             @JsonProperty("credits_remaining") Integer creditsRemaining,
-            @JsonProperty("id") String id,
+            @JsonProperty("id") UUID id,
             @JsonProperty("ip_address") IpAddress ipAddress,
             @JsonProperty("risk_score") Double riskScore,
             @JsonProperty("shipping_address") ShippingAddress shippingAddress,
@@ -27,6 +32,8 @@ public final class InsightsResponse extends ScoreResponse {
         super(creditsRemaining, id, riskScore, warnings);
         this.billingAddress = billingAddress == null ? new BillingAddress() : billingAddress;
         this.creditCard = creditCard == null ? new CreditCard() : creditCard;
+        this.device = device == null ? new Device() : device;
+        this.email = email == null ? new Email() : email;
         this.ipAddress = ipAddress == null ? new IpAddress() : ipAddress;
         this.shippingAddress = shippingAddress == null ? new ShippingAddress() : shippingAddress;
     }
@@ -45,6 +52,20 @@ public final class InsightsResponse extends ScoreResponse {
     @JsonProperty("credit_card")
     public CreditCard getCreditCard() {
         return creditCard;
+    }
+
+    /**
+     * @return The {@code Device} model object.
+     */
+    public Device getDevice() {
+        return device;
+    }
+
+    /**
+     * @return The {@code Email} model object.
+     */
+    public Email getEmail() {
+        return email;
     }
 
     /**
@@ -72,6 +93,8 @@ public final class InsightsResponse extends ScoreResponse {
         sb.append(", warnings=").append(this.warnings);
         sb.append(", ipAddress=").append(this.ipAddress);
         sb.append(", creditCard=").append(this.creditCard);
+        sb.append(", device=").append(this.device);
+        sb.append(", email=").append(this.email);
         sb.append(", shippingAddress=").append(this.shippingAddress);
         sb.append(", billingAddress=").append(this.billingAddress);
         sb.append('}');